car_web_frame/backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
pymysql==1.0.2
python-dotenv==0.19.0 